feat(tasks): add OpenInTabButton component with hover card for opening Pochi in a new tab
