+++ packages/vscode-webui/src/routes/tasks.tsx
@@ -2,6 +2,11 @@ import { Badge } from "@/components/ui/badge";
 import { Button } from "@/components/ui/button";
 import { Calendar } from "@/components/ui/calendar";
 import {
+  HoverCard,
+  HoverCardContent,
+  HoverCardTrigger,
+} from "@/components/ui/hover-card";
+import {
   Pagination,
   PaginationContent,
   PaginationEllipsis,
@@ -29,13 +34,18 @@ import {
   Edit3,
   GitBranch,
   HelpCircle,
+  SquareArrowOutUpRightIcon,
   TerminalIcon,
   Wrench,
   Zap,
 } from "lucide-react";
 import { useMemo, useState } from "react";
 import { MdOutlineErrorOutline } from "react-icons/md";
+import { useTranslation } from "react-i18next";
 import { useStoreDate } from "../livestore-provider";
+import { useSettingsStore } from "@/features/settings";
+import { vscodeHost } from "@/lib/vscode";
+import { useQuery } from "@tanstack/react-query";
 
 export const Route = createFileRoute("/tasks")({
   validateSearch: (search: Record<string, unknown>): { page?: number } => {
@@ -229,7 +239,7 @@ function Tasks() {
               </Pagination>
             </div>
           )}
-          <div className="w-24" />
+          <OpenInTabButton />
         </div>
       </div>
     </div>
@@ -353,6 +363,52 @@ function GitBadge({
       <GitBranch className="shrink-0" />
       <span className="truncate">{git.branch}</span>
     </Badge>
+  );
+}
+
+function OpenInTabButton() {
+  const { t } = useTranslation();
+  const { openInTab } = useSettingsStore();
+  const { data: credentials } = useQuery({
+    queryKey: ["pochi-credentials"],
+    queryFn: () => vscodeHost.readPochiCredentials(),
+  });
+
+  const handleOpenInTab = async () => {
+    await vscodeHost.openPochiInNewTab();
+  };
+
+  const showOpenInTab = openInTab && credentials;
+
+  if (!showOpenInTab) {
+    return <div className="w-6" />;
+  }
+
+  return (
+    <HoverCard>
+      <HoverCardTrigger asChild>
+        <span>
+          <Button
+            variant="ghost"
+            size="icon"
+            onClick={handleOpenInTab}
+            className="button-focus relative mr-2 h-6 w-6 p-0"
+          >
+            <span className="size-4">
+              <SquareArrowOutUpRightIcon className="size-4.5" />
+            </span>
+          </Button>
+        </span>
+      </HoverCardTrigger>
+      <HoverCardContent
+        side="top"
+        align="end"
+        sideOffset={6}
+        className="!w-auto max-w-sm bg-background px-3 py-1.5 text-xs"
+      >
+        {t("chat.openInTab")}
+      </HoverCardContent>
+    </HoverCard>
   );
 }
